Better handling of program startup
Before it was very annoying to restart awesome because it would always
launch the startup programs. Now this is handled by a bash script that
checks if there is already a process. It also has the benefit of
checking if the program is installed, which avoids logging errors on
some machines.