Add eza aliases
Note that distros seem to handle the transition in different ways.
Arch still uses exa as the command while openSUSE uses eza.
So for now we just alias exa to eza.
I think Arch will probably also switch in the future so this will
probably need to be reevaluated in the future.