Spacemacs: Add custom rust-analyzer path
Since Emacs is now supposed to run inside a container it somehow was
sometimes using non matching version of rust analyzer. I'm still not
sure where that mismatch was originating from, but this should ensure
that a up to date version is used.
Requires installing rustup inside my dev container.